feat: rename to generate-dts
