+++ package.json
@@ -13,15 +13,15 @@
     "setup:skip": "PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true yarn install && npm run build",
     "watch": "ts-node ./scripts/watch.ts",
     "build": "ts-node ./scripts/build.ts",
-    "bundle:dts": "ts-node ./scripts/bundle-dts.ts",
+    "generate:dts": "ts-node ./scripts/generate-dts.ts",
     "build:swc": "rm -rf packages/swc/native/** && napi build --platform --release --cargo-cwd packages/swc packages/swc/native",
     "version": "ts-node ./scripts/tag-version.ts && napi version -p packages/swc/npm -c packages/swc/package.json",
     "version:check": "ts-node ./scripts/version-check.ts",
-    "publish": "npm run copy:swc && ts-node ./scripts/publish-package.ts",
-    "publish:beta": "npm run copy:swc && PUBLISH_TYPE=beta ts-node ./scripts/publishPackageWithDistTag.ts",
-    "publish:next": "npm run copy:swc && PUBLISH_TYPE=next VERSION_PREFIX=rc ts-node ./scripts/publishPackageWithDistTag.ts",
-    "publish:stable": "PUBLISH_TAG=release-1 ts-node ./scripts/publish-package.ts",
-    "publish:stable-beta": "PUBLISH_TAG=release-1 ts-node ./scripts/publish-beta-package.ts",
+    "publish": "npm run copy:swc && npm run generate:dts && ts-node ./scripts/publish-package.ts",
+    "publish:beta": "npm run copy:swc && && npm run generate:dts PUBLISH_TYPE=beta ts-node ./scripts/publishPackageWithDistTag.ts",
+    "publish:next": "npm run copy:swc && && npm run generate:dts PUBLISH_TYPE=next VERSION_PREFIX=rc ts-node ./scripts/publishPackageWithDistTag.ts",
+    "publish:stable": "npm run generate:dts && PUBLISH_TAG=release-1 ts-node ./scripts/publish-package.ts",
+    "publish:stable-beta": "npm run generate:dts && PUBLISH_TAG=release-1 ts-node ./scripts/publish-beta-package.ts",
     "sync": "ts-node ./scripts/sync.ts",
     "rollback": "ts-node ./scripts/rollback.ts",
     "owner": "ts-node ./scripts/owner.ts",
